Added script to create clustered indices on tables. In the pullData script, now checks for status codes that aren't 200.

+++ bin/lib/3.pullDataFromCensusBureau.js
@@ -2,6 +2,7 @@
 /*
   command line args:
     --states: String with states to upload. States separated by spaces.
+    --release: QWI release to pull from. (Overrides value in config/qwi.env)
 */
 
 
@@ -23,10 +24,11 @@ const envFile = require('node-env-file')
 envFile(path.join(projectRoot, 'config/qwi.env'))
 const env = process.env
 
-const qwi_release = env.QWI_RELEASE
+let qwi_release = argv.release || env.QWI_RELEASE
 if (!qwi_release) {
   console.error("The QWI_RELEASE environment variable must be set in config/qwi.env")
 }
+qwi_release = qwi_release.trim()
 
 const dataDir = path.join(projectRoot, `data/metros_${qwi_release}`)
 mkdirp.sync(dataDir)
@@ -55,6 +57,8 @@ process.on('SIGINT', () => {
 
 let currentFilePath = null
 
+
+let filesWhereErrorEncountered = []
 let pullTheData = (state, workerC, firmC, cb) => {
 
     let tableName = `${workerC}_${firmC}_${aggregation}`
@@ -70,29 +74,43 @@ let pullTheData = (state, workerC, firmC, cb) => {
         return cb(null)
       }
 
-      currentFilePath = filePath
-      let fileStream = fs.createWriteStream(filePath) 
-
       let url = `${censusURL}/${state}/${qwi_release}/DVD-${workerC}_${firmC}/qwi_${state}_${fileName}`
 
-      http.request(url, res => {
-        res.pipe(fileStream)
-          .on('finish', () => {
-            currentFilePath = null
-            console.log("Downloaded", fileName, 'for', state + '.')
-            cb(null)
-          })
-          .on('error', (err) => {
-            console.error('\tERROR downloading the file')
-              fs.unlink(filePath, (fErr) => {
-                if (fErr) {
-                  console.error('\tCould not delete the file.')
-                  return cb(err)
-                }    
-                console.error('\tThe download file has been deleted.')
-              })
-            cb(err)
-          })
+      let timerLabel = `Downloaded ${fileName} for ${state}`
+      console.time(timerLabel)
+
+      return http.request(url, res => {
+        if (res.statusCode === 200) {
+
+          currentFilePath = filePath
+          let fileStream = fs.createWriteStream(filePath) 
+
+          res.pipe(fileStream)
+             .on('finish', () => {
+               currentFilePath = null
+               console.timeEnd(timerLabel)
+               cb(null)
+             })
+             .on('error', () => {
+               filesWhereErrorEncountered.push(filePath)
+               console.timeEnd(timerLabel)
+               console.error('\tERROR downloading the file.')
+               return fs.unlink(filePath, (fErr) => {
+                 if (fErr) {
+                   console.error('\tCould not delete the file.')
+                   console.error(fErr.stack)
+                 } else {
+                   console.error('\tThe download file has been deleted.')
+                 }   
+                 return cb(null) // async.series should continue despite error with a single file.
+               })
+             })
+        } else {
+          filesWhereErrorEncountered.push(filePath)
+          console.timeEnd(timerLabel)
+          console.error(`\tERROR: a status code of ${res.statusCode} was received.`)
+          cb(null) // async.series should continue despite error with a single file.
+        }
        }).end()
     }) 
 }
@@ -108,9 +126,11 @@ NOTE: Downloading the QWI data to ${dataDir}.
       It will be safe to delete that directory after the data is uploaded to Postgres.
       The pullDataFromCensusBureau script skips files already in that directory.
 `)
-console.time('Download all metro files.')
-async.series(scrapers, (err) => {
-  console.timeEnd('Download all metro files.')
 
-  if (err) { return console.error(err) }
+
+async.series(scrapers, () => {
+  if (filesWhereErrorEncountered.length) { 
+    console.error("Errors were encountered while downloading the following files:\n", 
+                  JSON.stringify(filesWhereErrorEncountered, null, 4))
+  }
 })
